+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - relay: add option relay.network.websocket_permessage_deflate ([#1549](https://github.com/weechat/weechat/issues/1549))
 - relay: add option relay.look.raw_messages_max_length ([#2122](https://github.com/weechat/weechat/issues/2122))
 - relay: add infos "relay_api_version" and "relay_api_version_number"
+- core: add completion "bars_items"
 
 ### Fixed